ver 0.1b
minor changes

+++ deggl-mandl.h
@@ -17,8 +17,8 @@ int buttonState = HIGH;
 
 // Allgemeine Variablen
 
-float current = 0;
-float currentAfterburner = 0;
+int current = 0;
+int currentAfterburner = 0;
 int torqCurrent = 1800;     // max. Strom in mA
 int afterburner = 80;       // Nachlauf in ms. Definiert das Drehmoment ! 
 int rpmPWM = 255;           // PWM für Motordrehzahl
